requirements.txt
Django==4.1.7
psycopg2==2.9.5
sentry_sdk==1.16.0
python-dotenv==1.0.0
wagtail==4.2
wagtail-meta-preview==1.1.1
wagtail_headless_preview==0.5.0

django-cors-headers


black
django-debug-toolbar
django-stubs==1.15.0
djangorestframework-stubs==1.9.1
isort
pipdeptree
ptvsd
ruff